pyproject.toml
[project]
name = "binge-buddy"
version = "0.1.0"
description = ""
authors = [
    {name = "shreyas",email = "shreyaskalvankar@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.9,<4.0"
dependencies = [
    "openai (>=1.65.4,<2.0.0)",
    "langchain (>=0.3.20,<0.4.0)",
    "langchain-openai (>=0.3.7,<0.4.0)",
    "openai-whisper (>=20240930,<20240931)",
    "setuptools-rust (>=1.10.2,<2.0.0)",
    "flask (>=3.1.0,<4.0.0)",
    "pymongo (>=4.11.2,<5.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "ollama (>=0.4.7,<0.5.0)",
    "transformers (>=4.49.0,<5.0.0)",
    "langgraph (==0.3.1)",
]

[tool.poetry]
packages = [{include = "binge_buddy", from = "src"}]


[tool.poetry.group.dev.dependencies]
pylint = "^3.3.4"
isort = "^6.0.1"
black = "^25.1.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
